Updating the web10g c measurement programs
Removed old web10g_request and web10g_receive and added web10g-stats.c
The new c program incoroporates the code from web10g_request and
receive and uses nonblocking io and the select statement to only
receive data when there is something to send.
